Revert __main__ move
